Physics and Timers
Added on the physic engine class for later implementation. Has some
basic equations for distance at the moment. Also added a button timer
class for later calculation of how long buttons are held down. We could
use this info to calculate velocities for later physics. Seems like the
screen needs to be bigger and more flexible. Should increase screen size
to window as well as build generic pathing of the character.